+++ src/IRremote.h
@@ -57,7 +57,6 @@
 /****************************************************
  *                     PROTOCOLS
  ****************************************************/
-
 /*
  * Supported IR protocols
  * Each protocol you include costs memory and, during decode, costs time
@@ -98,10 +97,10 @@
 
 //#define DEBUG // Activate this for lots of lovely debug output from the IRremote core.
 
-
 /****************************************************
  *                    RECEIVING
  ****************************************************/
+
 /**
  * MARK_EXCESS_MICROS is subtracted from all marks and added to all spaces before decoding,
  * to compensate for the signal forming of different IR receiver modules
@@ -175,7 +174,6 @@
  * Attention, active state of open drain is LOW, so connect the send LED between positive supply and send pin!
  */
 //#define USE_OPEN_DRAIN_OUTPUT_FOR_SEND_PIN
-
 /**
  * This amount is subtracted from the on-time of the pulses generated for software PWM generation.
  * It should be the time used for digitalWrite(sendPin, LOW) and the call to delayMicros()